docker/requirements_cmpsite.txt
asgiref==3.3.1
Django==3.1.6
django-health-check==3.16.2
djangorestframework==3.12.2
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
